+++ app/build.gradle.kts
@@ -14,7 +14,7 @@ android {
 
     defaultConfig {
         applicationId = "com.skgtecnologia.sisem"
-        minSdk = 31
+        minSdk = 30
         targetSdk = 34
         versionCode = 2
         versionName = "0.2.0"
