+++ app/src/main/java/com/skgtecnologia/sisem/ui/menu/MenuDrawer.kt
@@ -1,20 +1,11 @@
 package com.skgtecnologia.sisem.ui.menu
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Menu
 import androidx.compose.material3.DrawerState
 import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -23,21 +14,17 @@ import com.skgtecnologia.sisem.ui.menu.items.getDrawerMenuItemList
 import com.skgtecnologia.sisem.ui.navigation.MainNavigationRoute
 import com.valkiria.uicomponents.components.banner.OnErrorHandler
 import com.valkiria.uicomponents.components.loader.LoaderComponent
-import com.valkiria.uicomponents.components.map.MapComponent
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-@Suppress("LongMethod")
+@Suppress("LongMethod", "LongParameterList")
 @Composable
 fun MenuDrawer(
-    modifier: Modifier = Modifier,
+    drawerState: DrawerState,
     onClick: (MainNavigationRoute) -> Unit,
     onLogout: () -> Unit,
     content: @Composable () -> Unit
 ) {
     val viewModel = hiltViewModel<MenuViewModel>()
-    val coroutineScope: CoroutineScope = rememberCoroutineScope()
-    val drawerState: DrawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
 
     val uiState = viewModel.uiState
 
@@ -71,28 +58,7 @@ fun MenuDrawer(
         drawerState = drawerState,
         gesturesEnabled = drawerState.isOpen
     ) {
-        Box(
-            modifier = modifier.fillMaxSize(),
-        ) {
-            MapComponent(
-                modifier = Modifier.fillMaxSize(),
-                coordinates = -75.5657751 to 6.2082622
-            )
-
-            IconButton(
-                onClick = {
-                    coroutineScope.launch {
-                        drawerState.open()
-                    }
-                }
-            ) {
-                Icon(
-                    imageVector = Icons.Outlined.Menu,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.primary
-                )
-            }
-        }
+        content()
     }
 
     OnErrorHandler(uiModel = uiState.errorModel) {
@@ -108,9 +74,8 @@ fun MenuDrawer(
 @Composable
 fun MenuDrawerPreview() {
     MenuDrawer(
+        rememberDrawerState(initialValue = DrawerValue.Closed),
         onClick = {},
         onLogout = {}
-    ) {
-
-    }
+    ) {}
 }
